synq: extract DDL reader, split CTE handling, methodize catalog helpers
catalog.rs and analyzer/pass.rs were doing too much, and a lot of
DDL-shaped logic lived as standalone free functions threading
(stmt, roles) tuples through the call sites. Pull those concerns
apart into focused modules and convert the helpers into methods
on the types that own the data.

- semantic/ddl.rs (new): DdlReader<'a, 'stmt> bundles
  (stmt, roles) and exposes name_span / column_spans /
  extract_columns / columns_from_select / expr_source_text /
  is_table_returning / function_arity / span_field_text.
  Catalog::accumulate_ddl, record_ddl_definition_site,
  analyzer/mod.rs, analyzer/pass.rs, and lineage/resolver.rs all
  call into the reader instead of scattered free functions.
- catalog.rs (1614 -> 1176 lines): removed canonical_name (inlined),
  overload_accepts -> FunctionOverload::accepts,
  expected_fixed_arities -> FunctionSet::fixed_arities,
  map_function_category -> From<DialectFunctionCategory>,
  build_dialect_layer -> CatalogLayerContents::populate_dialect_builtins,
  push_unique folded into shared Catalog::unique_names_across_layers
  helper used by all three all_*_names methods.
- analyzer/pass/{mod.rs (684), cte.rs (301)}: visit_cte_scope and
  its helpers (extract_cte_binding, extract_declared_cols,
  check_cte_column_count, count_result_columns,
  record_select_column_offsets, CteBindingInfo) moved to a sibling
  impl block in pass/cte.rs. pass/mod.rs keeps the dispatcher and
  the SourceRef / Call / ColumnRef / Query / ScopedSource / Trigger
  visitors.
- analyzer/helpers.rs: deleted ddl_name_offset (now
  DdlReader::name_span).

Net: -755 lines from existing files, +~681 in two focused new
files. 275 unit tests + ast/fmt/semantic/validate/lineage/lsp/
perfetto-fmt/perfetto-val integration suites all pass; clippy
clean across the workspace.
